refactor: clean up ambient functions and integrate with morph-detail.js
