+++ src/pages/downloads.astro
@@ -85,8 +85,6 @@ import {
 
       <article
         id="downloads-active-list"
-        role="status"
-        aria-live="polite"
         aria-label="Active downloads"
         class="rounded-2xl border border-line bg-surface overflow-hidden flex flex-col divide-y divide-line/60">
       </article>
